requirements.txt
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pygments==2.10.0
python-dotenv==0.19.0
rich==10.11.0
Werkzeug==2.0.1
